requirements.txt
pydantic~=2.12.4
PyYAML~=6.0.3
pydantic-settings~=2.12.0
cachetools~=6.2.3
dependency-injector @ git+https://github.com/pavalso/python-dependency-injector
pyOpenSSL~=25.3.0
cryptography~=46.0.3
filelock~=3.20.0
packaging~=25.0
platformdirs~=4.5.1